gemini/workers/geo/requirements.txt
rasterio>=1.3,<2
Pillow>=10,<11
numpy>=1.24,<2
minio>=7.2,<8
redis>=5,<6
requests>=2.31,<3
